requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-openai
pydantic
qdrant-client
openai
python-dotenv
pytest
playwright
rich
faiss-cpu
